+++ packages/errors/package.json
@@ -5,7 +5,7 @@
   "keywords": [],
   "author": "Endo contributors",
   "license": "Apache-2.0",
-  "homepage": "https://github.com/endojs/endo/tree/master/packages/skel#readme",
+  "homepage": "https://github.com/endojs/endo/tree/master/packages/errors#readme",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/endojs/endo.git"
