+++ packages/exo/package.json
@@ -32,12 +32,12 @@
     "test": "ava"
   },
   "dependencies": {
+    "@endo/common": "^1.0.1",
     "@endo/env-options": "^1.0.1",
     "@endo/eventual-send": "^1.0.1",
     "@endo/far": "^1.0.1",
     "@endo/pass-style": "^1.0.1",
-    "@endo/patterns": "^1.0.1",
-    "@endo/utils": "^1.0.1"
+    "@endo/patterns": "^1.0.1"
   },
   "devDependencies": {
     "@endo/init": "^1.0.1",
