+++ packages/patterns/src/keys/checkKey.js
@@ -12,7 +12,7 @@ import {
   makeFullOrderComparatorKit,
   sortByRank,
 } from '@endo/marshal';
-import { identChecker } from '@endo/utils';
+import { identChecker } from '@endo/common';
 
 import { checkElements, makeSetOfElements } from './copySet.js';
 import { checkBagEntries, makeBagOfEntries } from './copyBag.js';
