--- packages/utils/README.md
@@ -1,1 +1,0 @@
-
