--- packages/utils/test/test-index.js
@@ -1,5 +1,0 @@
-import { test } from './prepare-test-env-ava.js';
-
-test('place holder', async t => {
-  t.pass();
-});
